add VraNgPolicyBase class
