+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/helpers/filesystem/ContentSharingPolicyFsMocks.java
@@ -17,7 +17,7 @@ package com.vmware.pscoe.iac.artifact.helpers.filesystem;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.vmware.pscoe.iac.artifact.model.vrang.VraNgContentSharingPolicy;
+import com.vmware.pscoe.iac.artifact.model.vrang.ariaPolicies.VraNgContentSharingPolicy;
 import java.io.File;
 import java.nio.file.Paths;
 import java.nio.file.Path;
